+++ packages/analytics/analytics-chart/src/components/chart-types/TimeSeriesChart.vue
@@ -59,6 +59,7 @@ import { reactive, ref, computed, toRef, inject, watch, onUnmounted, useTemplate
 import 'chartjs-adapter-date-fns'
 import 'chart.js/auto'
 import { VerticalLinePlugin } from '../chart-plugins/VerticalLinePlugin'
+import { CoordinatorPlugin } from '../chart-plugins/CoordinatorPlugin'
 import { HighlightPlugin } from '../chart-plugins/HighlightPlugin'
 import { DragSelectPlugin } from '../chart-plugins/DragSelectPlugin'
 import type { DragSelectEventDetail } from '../chart-plugins/DragSelectPlugin'
@@ -66,14 +67,15 @@ import ToolTip from '../chart-plugins/ChartTooltip.vue'
 import ChartLegend from '../chart-plugins/ChartLegend.vue'
 import { Line, Bar } from 'vue-chartjs'
 import composables from '../../composables'
-import type { Threshold, TooltipState, ZoomActionItem } from '../../types'
+import type { Threshold, TooltipState, YAxisConfig, ZoomActionItem } from '../../types'
 import { type ChartLegendSortFn, type ChartTooltipSortFn, type EnhancedLegendItem, type KChartData, type LegendValues, type TooltipEntry } from '../../types'
-import type { GranularityValues, AbsoluteTimeRangeV4, ExploreAggregations } from '@kong-ui-public/analytics-utilities'
+import type { GranularityValues, AbsoluteTimeRangeV4, ExploreAggregations, InteractionCoordinator } from '@kong-ui-public/analytics-utilities'
 import type { Chart, Plugin } from 'chart.js'
 import { ChartLegendPosition } from '../../enums'
 import { generateLegendItems } from '../../utils'
 import { hasExactlyOneDatapoint } from '../../utils/commonOptions'
 import { ThresholdPlugin } from '../chart-plugins/ThresholdPlugin'
+import { INJECT_DASHBOARD_COORDINATOR } from '../../constants'
 
 interface TimeSeriesChartProps {
   chartData?: KChartData
@@ -94,6 +96,8 @@ interface TimeSeriesChartProps {
   zoomActionItems?: ZoomActionItem[]
   tooltipMetricDisplay?: string
   threshold?: Record<ExploreAggregations, Threshold[]>
+  leftYAxisGrid?: boolean
+  rightYAxis?: YAxisConfig
 }
 
 const props = withDefaults(
@@ -115,6 +119,8 @@ const props = withDefaults(
     zoomActionItems: undefined,
     tooltipMetricDisplay: '',
     threshold: undefined,
+    leftYAxisGrid: undefined,
+    rightYAxis: undefined,
   },
 )
 
@@ -137,6 +143,13 @@ const legendPosition = inject('legendPosition', ChartLegendPosition.Bottom)
 const chartParentRef = useTemplateRef<HTMLDivElement>('chartParent')
 const zoomTimeRange = ref<AbsoluteTimeRangeV4 | undefined>(undefined)
 const isDoingSelection = ref(false)
+const coordinator: InteractionCoordinator | null = inject(INJECT_DASHBOARD_COORDINATOR, null)
+const coordinatorPlugin = new CoordinatorPlugin({
+  coordinator,
+  triggerOnSelf: true,
+  watchTimestamp: true,
+  watchDimension: true,
+})
 
 const tooltipData: TooltipState = reactive({
   showTooltip: false,
@@ -170,6 +183,7 @@ const htmlLegendPlugin: Plugin = {
 }
 
 const plugins = computed(() => [
+  coordinatorPlugin,
   htmlLegendPlugin,
   highlightPlugin,
   ...(props.brush ? [dragSelectPlugin] : []),
@@ -193,6 +207,8 @@ const { options } = composables.useLineChartOptions({
   dimensionAxesTitle: toRef(props, 'dimensionAxesTitle'),
   pointsWithoutHover: pointsWithoutHover,
   threshold: toRef(props, 'threshold'),
+  leftYAxisGrid: toRef(props, 'leftYAxisGrid'),
+  rightYAxis: toRef(props, 'rightYAxis'),
 })
 
 composables.useReportChartDataForSynthetics(toRef(props, 'chartData'), toRef(props, 'syntheticsDataKey'))